Fixed syntax issues removed use of interface

+++ src/start/Main.java
@@ -1,10 +1,10 @@
 package start;
 
-import genetic.*;
-import creatures.*;
+
+import evolution.CharEvolution;
 
 public class Main {
   public static void main(String[] args) {
-    CharEvolution evolution = new CharEvolution (50, 0.10, "Hello World");
+    CharEvolution evolution = new CharEvolution (50, (float)0.10, "Hello World");
   }
 }
